// tsundoku/blueprints/ux/static/ts/api/types.gen.ts
export type TorrentConfigResponse = {
  /**
   * Client
   */
  client: string;
  /**
   * Host
   */
  host: string;
  /**
   * Port
   */
  port: number;
  /**
   * Username
   */
  username: string | null;
  /**
   * Password
   */
  password: string | null;
  /**
   * Secure
   */
  secure: boolean;
};

/**
 * TorrentConfigUpdate
 */
export type TorrentConfigUpdate = {
  /**
   * Client
   */
  client?: string | null;
  /**
   * Host
   */
  host?: string | null;
  /**
   * Port
   */
  port?: number | null;
  /**
   * Username
   */
  username?: string | null;
  /**
   * Password
   */
  password?: string | null;
  /**
   * Secure
   */
  secure?: boolean | null;
};

/**
 * TorrentTestResult
 */
export type TorrentTestResult = {
  /**
   * Success
   */
  success: boolean;
  /**
   * Error
   */
  error?: string | null;
};